added a proof of concept gui.. with node drawing and zooming

+++ tagcwebapp/src/main/java/com/pl/tagc/tagcwebapp/NodeService.java
@@ -20,14 +20,14 @@ public class NodeService {
 	// The Java method will produce content identified by the MIME Media
 	// type "application/json"
 	@Produces("application/json")
-	public ResultObject smooth(@DefaultValue("0") @QueryParam("xleft") int xleft,
-			@DefaultValue("0") @QueryParam("ytop") int ytop, @DefaultValue("100") @QueryParam("xright") int xright,
-			@DefaultValue("100") @QueryParam("ybtm") int ybtm) {
+	public ResultObject smooth(@DefaultValue("0") @QueryParam("xleft") double xleft,
+			@DefaultValue("0") @QueryParam("ytop") double ytop, @DefaultValue("100") @QueryParam("xright") double xright,
+			@DefaultValue("100") @QueryParam("ybtm") double ybtm) {
 		ResultObject r = getNodes(xleft, ytop, xright, ybtm);
 		return r;
 	}
 
-	private ResultObject getNodes(int xleft, int ytop, int xright, int ybtm) {
+	private ResultObject getNodes(double xleft, double ytop, double xright, double ybtm) {
 		System.out.println(ytop);
 		System.out.println(xleft);
 		System.out.println(xright);
